requirements.txt
anyio==3.6.2
attrs==22.2.0
autopep8==2.0.1
certifi==2022.12.7
click==8.1.3
cycler==0.11.0
exceptiongroup==1.1.0
fastapi==0.89.0
fonttools==4.38.0
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
importlib-metadata==6.0.0
iniconfig==2.0.0
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.1
packaging==22.0
Pillow==9.4.0
pluggy==1.0.0
pycodestyle==2.10.0
pydantic==1.10.4
pydicom==2.3.1
pyparsing==3.0.9
pytest==7.2.0
python-dateutil==2.8.2
python-dotenv==0.21.0
PyYAML==6.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
tomli==2.0.1
typing_extensions==4.4.0
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
zipp==3.11.0
